优化代码结构

+++ FUCK-GUET-WiFi.py
@@ -6,10 +6,13 @@ import os
 import getpass
 import subprocess
 import time
+import sys
 
 
 def get_wifi_info():
-    output = subprocess.run(["netsh", "WLAN", "show", "interfaces"], capture_output=True, text=True, shell=True)
+    output = subprocess.run(["netsh", "WLAN", "show", "interfaces"],
+                            capture_output=True,
+                            text=True, shell=True)
     output = output.stdout.replace(" ", "").split()
     tmp_list = {}
     for line in output:
@@ -21,26 +24,44 @@ def get_wifi_info():
     return tmp_list
 
 
+connect_retry_times = 0
+
 wifi_info = get_wifi_info()
 mac = wifi_info['物理地址']
-if "SSID" not in wifi_info:
-    print("WiFi未连接,正在尝试连接...")
-    try:
-        os.system("netsh WLAN connect GUET-WiFi")
-        time.sleep(3)  # 等待wifi连接
-    except:
-        print("连接失败!请手动连接或稍后再试。")
-        exit(0)
-elif wifi_info["SSID"] != "GUET-WiFi":
-    print("未正确连接至校园网,正在尝试连接...")
-    try:
-        os.system("netsh WLAN disconnect")
-        time.sleep(1)
-        os.system("netsh WLAN connect GUET-WiFi")
-        time.sleep(3)
-    except:
-        print("连接失败!请手动连接或稍后再试。")
-        exit(0)
+
+while connect_retry_times <= 3:
+    if "SSID" not in wifi_info:
+        print("WiFi未连接,正在尝试连接...")
+        try:
+            subprocess.run(["netsh", "WLAN", "connect", "GUET-WiFi"],
+                           capture_output=False, stdout=subprocess.DEVNULL)
+            time.sleep(3)  # 等待wifi连接
+        except:
+            print("连接失败!请手动连接或稍后再试。")
+            os.system("pause")
+            sys.exit(0)
+    elif wifi_info["SSID"] != "GUET-WiFi":
+        print("未正确连接至校园网,正在尝试连接...")
+        try:
+            subprocess.run(["netsh", "WLAN", "disconnect"],
+                           capture_output=False, stdout=subprocess.DEVNULL)
+            time.sleep(1)  # 等待断开连接
+            subprocess.run(["netsh", "WLAN", "connect", "GUET-WiFi"],
+                           capture_output=False, stdout=subprocess.DEVNULL)
+            time.sleep(3)  # 等待重新连接
+        except:
+            print("连接失败!请手动连接或稍后再试。")
+            os.system("pause")
+            sys.exit(0)
+    else:
+        break
+    wifi_info = get_wifi_info()  # 重新尝试获取wifi信息
+    connect_retry_times += 1
+
+if "SSID" not in wifi_info or wifi_info["SSID"] != "GUET-WiFi":
+    print("wifi连接异常!请检查wifi连接、确保在校园网覆盖范围内或稍后再试")
+    os.system("pause")
+    sys.exit(0)
 
 if os.path.exists('info.json'):  # 文件存储
     print("正在使用已保存的账户密码登录...")
@@ -48,7 +69,6 @@ if os.path.exists('info.json'):  # 文件存储
         data = json.load(f)
         user_name = str(data.get('user_name'))
         pwd = str(data.get('pwd'))
-        f.close()
 else:  # 找不到文件时需用户输入
     user_name = str(input('首次使用,请输入智慧校园账号/学号:'))
     while len(user_name) != 10:
@@ -60,9 +80,11 @@ else:  # 找不到文件时需用户输入
     pwd = base64.b64encode(pwd.encode("ascii")).decode("ascii")  # base64加密
 
 # 生成登录url、header
-url = f"http://10.0.1.5:801/eportal/portal/login?user_account=%2C0%2C{user_name}&user_password={pwd}&wlan_user_mac={mac}&wlan_ac_name=HJ-BRAS-ME60-01"
+url = f"http://10.0.1.5:801/eportal/portal/login?user_account=%2C0%2C{user_name}\
+        &user_password={pwd}&wlan_user_mac={mac}&wlan_ac_name=HJ-BRAS-ME60-01"
 headers = {
-    'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/127.0.0.0 Safari/537.36 Edg/127.0.0.0',
+    'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 \
+    (KHTML, like Gecko) Chrome/127.0.0.0 Safari/537.36 Edg/127.0.0.0',
     'Referer': 'https://10.0.1.5:801/',
     'Host': '10.0.1.5'
 }
@@ -77,12 +99,16 @@ try:
         with open('info.json', 'w') as f:
             data = {'user_name': user_name, 'pwd': pwd}
             json.dump(data, f, indent=4)
-            f.close()
             print("已将账户密码加密存入程序目录下的info.json文件,下次认证无需输入")
-    elif "ldap auth error" in content:  # 如果密码错误,删除账户文件,以防下次再使用错误的账户密码
+
+    elif "ldap auth error" in content and os.path.exists('info.json'):
+        # 如果密码错误且账户文件存在,删除账户文件,以防下次再使用错误的账户密码
         os.remove('info.json')
-except Exception as ex:  # get超时3s后捕获
-    print(ex)
+
+except:  # get超时3s后捕获
     print("连接失败,可能是请求频率过高,请稍后再试...")
+    os.system("pause")
+    sys.exit(0)
 
 os.system("pause")
+sys.exit(0)
